#66 [fix] 해당 API response 값 변경에 따라 수정

+++ client/app/src/main/java/com/kaer/menuw/domain/entity/RecommendMenu.kt
@@ -5,8 +5,8 @@ data class RecommendMenu (
     val menuName: String,
     val menuImgUrl: String,
     val ingredients: String,
-    val cal: Int,
-    val na: Int,
+    val cal: Double,
+    val na: Double,
     val similarity: Double,
     val recommend: Boolean
 )
